Almost finished

+++ src/main/java/com/kostis/teachersdemo/repo/StudentCourseAssociationRepository.java
@@ -3,6 +3,6 @@ package com.kostis.teachersdemo.repo;
 import com.kostis.teachersdemo.entities.StudentCourseAssociation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudentCourseAssociationRepository extends JpaRepository<StudentCourseAssociation, Long>{
+public interface StudentCourseAssociationRepository extends JpaRepository<StudentCourseAssociation, Integer>{
     StudentCourseAssociation findByStudent_IdAndCourse_Id(Integer studentId, Integer courseId);
 }
